requirements.txt
git+https://github.com/facebookresearch/segment-anything.git
gradio==3.39
progressbar2
gdown
gitpython
hickle
tensorboard
numpy~=1.24.4
opencv-python~=4.7.0.72
matplotlib~=3.7.1
pyyaml
av
openmim
tqdm~=4.65.0
psutil
auth0-python==4.3.0
bandit==1.7.5
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
coverage==7.2.7
cryptography==41.0.1
cycler==0.10.0
datadog-api-client==2.12.0
debugpy==1.6.7
deprecated==1.2.14
dill==0.3.6
distlib==0.3.6
exceptiongroup==1.1.1
fal-serverless==0.6.35
filelock==3.12.2
fire==0.5.0
flake8==6.0.0
fonttools==4.40.0
gitdb==4.0.10
gitpython==3.1.31
grpc-interceptor==0.15.2
grpcio==1.56.0
identify==2.5.24
idna==2.10
importlib-metadata==6.0.1
importlib-resources==5.12.0
iniconfig==2.0.0
isolate-proto==0.0.31
isolate[build]==0.12.2
isort==5.12.0
jinja2==3.1.2
kiwisolver==1.4.4
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
mypy==1.4.1
networkx==3.1
nodeenv==1.8.0
numpy==1.24.4
opencv-python==4.7.0.72
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
packaging==23.1
parameterized==0.9.0
pathspec==0.11.1
pbr==5.11.1
pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
portalocker==2.7.0
pre-commit==3.3.3
protobuf==4.23.3
pybboxes==0.1.6
pycocotools~=2.0.7
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pygments==2.15.1
pyjwt[crypto]==2.7.0
pyparsing==2.4.7
pytest==7.4.0
python-dateutil==2.8.2
pytoolconfig[global]==1.2.5
pyupgrade==3.7.0
pyyaml==6.0
requests==2.31.0
rich==13.4.2
rope==1.8.0
ruff==0.0.244
sahi==0.11.14
setuptools==67.8.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
stevedore==5.1.0
structlog==22.3.0
sympy==1.12
tblib==2.0.0
termcolor==2.3.0
terminaltables==3.1.10
tokenize-rt==5.1.0
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
urllib3==1.26.16
virtualenv==20.23.1
wheel==0.38.4
wrapt==1.15.0
zipp==3.15.0
metric-temporal-logic